Casi terminado

+++ BetPlay/Masajista.cs
@@ -5,7 +5,7 @@ namespace BetPlay
         public string titulacion { get; set; }
         public int añosExp { get; set; }
 
-        public Masajista(string Nombre, int Edad, string CiudadOrigen, long Id, string Titulacion,int AñosExp) : base(Nombre,Edad,CiudadOrigen,Id){
+        public Masajista(string Nombre, int Edad, string CiudadOrigen, int Id, string Titulacion,int AñosExp) : base(Nombre,Edad,CiudadOrigen,Id){
             this.titulacion = Titulacion;
             this.añosExp = AñosExp;
         }
